00:00 7/4/21
New playerField.cpp .h
Rearrange global variable
