+++ .gitignore
@@ -1,7 +1,10 @@
 
 .vs/
+.x64\Debug
+*.vcxproj
 *.exe
 *.ilk
 *.log
 *.pdb
 *.tlog
+*.obj
